+++ app/controllers/projects_controller.rb
@@ -138,7 +138,9 @@ class ProjectsController < ApplicationController
         :name, :description, :emails, :acrostic_enabled, :subject_code_name,
         :show_contacts, :show_documents, :show_posts,
         # Uploaded Logo
-        :logo, :logo_uploaded_at, :logo_cache
+        :logo, :logo_uploaded_at, :logo_cache,
+        # Will automatically generate a site if the project has no site
+        :site_name
       )
     end